stmgr: Put actor state accessors in actors.go
